+++ test/unit/capacityEngine.js
@@ -1,63 +1,10 @@
 const sinon = require('sinon');
 const { expect } = require('chai');
-const capacityEngine = require('../../src/lib/capacityEngine');
-const mockStore = require('../mocks/store');
 const { BigNumber } = require('ethers');
 
-const capacities = [
-  {
-    productId: 0,
-    capacity: [
-      {
-        assetId: 0,
-        amount: '4761714669056628480',
-      },
-      {
-        assetId: 1,
-        amount: '13305160151201388636532',
-      },
-      {
-        assetId: 255,
-        amount: '463200000000000000000',
-      },
-    ],
-  },
-  {
-    productId: 1,
-    capacity: [
-      {
-        assetId: 0,
-        amount: '2380857334528314240',
-      },
-      {
-        assetId: 1,
-        amount: '6652580075600694318266',
-      },
-      {
-        assetId: 255,
-        amount: '231600000000000000000',
-      },
-    ],
-  },
-  {
-    productId: 2,
-    capacity: [
-      {
-        assetId: 0,
-        amount: '4761714669056628480',
-      },
-      {
-        assetId: 1,
-        amount: '13305160151201388636532',
-      },
-      {
-        assetId: 255,
-        amount: '463200000000000000000',
-      },
-    ],
-  },
-];
-
+const capacityEngine = require('../../src/lib/capacityEngine');
+const mockStore = require('../mocks/store');
+const { capacities } = require('./responses');
 describe('Capacity Engine tests', () => {
   const store = { getState: () => null };
 
@@ -68,7 +15,6 @@ describe('Capacity Engine tests', () => {
   it('should return capacity for all products', () => {
     sinon.stub(store, 'getState').callsFake(() => mockStore);
     const now = BigNumber.from(Date.now()).div(1000);
-
     const response = capacityEngine(store, [], now);
 
     response.forEach((product, i) => {
